ml_sample/arxiv-importer-v2/requirements.txt
google-generativeai
pypdf
torch
requests
ruff
mypy
pytest
types-requests
